fix: percentage should be float

+++ main.py
@@ -8,4 +8,4 @@ config.read('config.ini')
 
 adapter=AdapterMapshaper2Simplifier()
 simplifier=Simplifier(input_file=str(config['simplifier']['input']), adapter=adapter)
-boundary_simplified=simplifier.simplify(int(config['simplifier']['percentage']))
+boundary_simplified=simplifier.simplify(float(config['simplifier']['percentage']))
